is it working ?

+++ include/PPN-microbench/microbench.hpp
@@ -16,16 +16,16 @@ using json = nlohmann::ordered_json;
 class Microbench {
   protected:
     std::string name;
-    int nbIterations;
+    u64 nbIterations;
 
   public:
     Context context = Context::getInstance();
 
-    Microbench(std::string name, int nbIterations);
+    Microbench(std::string name, u64 nbIterations);
     virtual ~Microbench();
 
     std::string getName();
-    int getNbIterations();
+    u64 getNbIterations();
 
     virtual void run() = 0;
     virtual json getJson() = 0;
