+++ src/PPN-microbench/microbench.cpp
@@ -1,6 +1,6 @@
 #include <PPN-microbench/microbench.hpp>
 
-Microbench::Microbench(std::string name, int nbIterations) {
+Microbench::Microbench(std::string name, u64 nbIterations) {
     this->name = name;
     this->nbIterations = nbIterations;
 }
@@ -9,4 +9,4 @@ Microbench::~Microbench() {}
 
 std::string Microbench::getName() { return this->name; }
 
-int Microbench::getNbIterations() { return this->nbIterations; }
+u64 Microbench::getNbIterations() { return this->nbIterations; }
